✨ feat: 출석 화면 screen,  timer, input UI 및 로직 구현
